% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{AllEdgeIntensities.intensitynet}
\alias{AllEdgeIntensities.intensitynet}
\title{Calculate all the edge intensities of the graph. It's more accurate than using iteratively the 
function EdgeIntensity for all edges.}
\usage{
\method{AllEdgeIntensities}{intensitynet}(obj, z = 5)
}
\arguments{
\item{obj}{intensitynet object}

\item{z}{Maximum distance between the event and the edge to consider the event part of the edge.}
}
\value{
intensitynet class object where the graph contains all the edge intensities as an attribute
}
\description{
Calculate all the edge intensities of the graph. It's more accurate than using iteratively the 
function EdgeIntensity for all edges.
}
