% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{AreEventsRelated}
\alias{AreEventsRelated}
\alias{AreEventsRelated.intensitynet}
\title{Checks if events are related to the intensitynet object}
\usage{
AreEventsRelated(obj)

\method{AreEventsRelated}{intensitynet}(obj)
}
\arguments{
\item{obj}{Intensitynet object}
}
\value{
TRUE if related, FALSE otherwise
}
\description{
Checks if events are related to the intensitynet object
}
\examples{

data("und_intnet_chicago")
AreEventsRelated(und_intnet_chicago)

}
