% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{CalculateDistancesMtx.netTools}
\alias{CalculateDistancesMtx.netTools}
\title{Calculates the distances between all pairs of nodes from the given network}
\usage{
\method{CalculateDistancesMtx}{netTools}(obj)
}
\arguments{
\item{obj}{netTools object -> list(): with the node coordinates 'x' and 'y'}
}
\value{
distances matrix
}
\description{
Calculates the distances between all pairs of nodes from the given network
}
