% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetDir.R
\name{CalculateEventIntensities.intensitynetDir}
\alias{CalculateEventIntensities.intensitynetDir}
\title{Calculates edgewise and mean nodewise intensities for Directed networks}
\usage{
\method{CalculateEventIntensities}{intensitynetDir}(obj)
}
\arguments{
\item{obj}{intensitynetDir object}
}
\value{
intensitynetDir object with a graph containing all the intensities as attributes of its nodes and edges
}
\description{
Calculates edgewise and mean nodewise intensities for Directed networks
}
