% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetMix.R
\name{CalculateEventIntensities.intensitynetMix}
\alias{CalculateEventIntensities.intensitynetMix}
\title{Calculates edgewise and mean nodewise intensities for Mixed networks}
\usage{
\method{CalculateEventIntensities}{intensitynetMix}(obj)
}
\arguments{
\item{obj}{intensitynetMix object}
}
\value{
intensitynetMix object with a graph containing all the intensities as attributes of its nodes and edges
}
\description{
Calculates edgewise and mean nodewise intensities for Mixed networks
}
