% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetUnd.R
\name{CalculateEventIntensities.intensitynetUnd}
\alias{CalculateEventIntensities.intensitynetUnd}
\title{Calculates edgewise and mean nodewise intensities for Undirected networks}
\usage{
\method{CalculateEventIntensities}{intensitynetUnd}(obj)
}
\arguments{
\item{obj}{intensitynetUnd object}
}
\value{
intensitynetUnd object with a graph containing all the intensities as attributes of its nodes and edges
}
\description{
Calculates edgewise and mean nodewise intensities for Undirected networks
}
