% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Castellon.R
\docType{data}
\name{Castellon}
\alias{Castellon}
\title{Traffic network for Castellon}
\format{An object of class \code{data.frame} with 1611 rows and 1611 columns.}
\usage{
data(Castellon)
}
\description{
Network for roadd map of Castellon stored as adjacenvy matrix
}
\references{
Eckardt, M and Mateu, J (2016). Point patterns occurring on complex structures in space and space-time: An alternative network approach.
}
\keyword{datasets}

