% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimes.R
\docType{data}
\name{crimes}
\alias{crimes}
\title{Artificial data on crimes committed on the traffic network of Castellon}
\format{An object of class \code{data.frame} with 5089 rows and 2 columns.}
\usage{
data(crimes)
}
\description{
Data provides the artificial coordinates of crimes commited on Castellon traffic network
}
\references{
Eckardt, M and Mateu, J (2016). Point patterns occurring on complex structures in space and space-time: An alternative network approach.
}
\keyword{datasets}

