% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dir_intnet_chicago}
\alias{dir_intnet_chicago}
\title{This data is an intensitynet object containing a directed network. 
The base data used is from Chicago, extracted from the spatstat package.}
\format{
An object of class \code{intensitynet} (inherits from \code{intensitynetDir}) of length 4.
}
\source{
\url{https://rdrr.io/cran/spatstat.data/man/chicago.html}
}
\usage{
dir_intnet_chicago
}
\description{
This data is an intensitynet object containing a directed network. 
The base data used is from Chicago, extracted from the spatstat package.
}
\keyword{datasets}
