% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{EdgeIntensity.intensitynet}
\alias{EdgeIntensity.intensitynet}
\title{Given two nodes, calculates its edge intensity}
\usage{
\method{EdgeIntensity}{intensitynet}(obj, node_id1, node_id2)
}
\arguments{
\item{obj}{intensitynet object}

\item{node_id1}{First node ID of the edge}

\item{node_id2}{Second node ID of the edge}
}
\value{
Intensity of the edge
}
\description{
If not calculated, calculates the intensity of the edge with nodes; node_id1, node_id2. 
If the edge already contains an intensity, the function gives it directly without re-calculation.
}
