% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{EdgeIntensity.intensitynetUnd}
\alias{EdgeIntensity.intensitynetUnd}
\alias{EdgeIntensity.intensitynet}
\title{If not calculated, calculates the intesnity of the edge with nodes; node_id1, node_id2. 
If the edge already contains an intensity, gives it directly.}
\usage{
\method{EdgeIntensity}{intensitynet}(obj, node_id1, node_id2, z = 5)
}
\arguments{
\item{node_id1}{First node ID of the edge}

\item{node_id2}{Second node ID of the edge}
}
\value{
edge_intensity - Intensity of the edge
}
\description{
If not calculated, calculates the intesnity of the edge with nodes; node_id1, node_id2. 
If the edge already contains an intensity, gives it directly.
}
