% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{GeoreferencedGgplot2.netTools}
\alias{GeoreferencedGgplot2.netTools}
\title{This function uses 'ggplot' to plot heatmaps of a network}
\usage{
\method{GeoreferencedGgplot2}{netTools}(obj, ...)
}
\arguments{
\item{obj}{netTools object -> list(
intnet: intensitynet object, 
data_df: dataframe(
  xcoord: x coordinates of the nodes, 
  ycoord: y coordinates of the nodes,
  value: vector values to plot
 ), 
net_vertices: chosen vertices to plot the heatmap (or its related edges in case to plot the edge heatmap),
heat_type: data which the heatmap will refer,
mode: ('moran', 'getis', 'v_intensity', 'e_intensity' or mark), 
show_events: boolean to show or not the events as orange squares 
)}

\item{...}{extra arguments for the ggplot}
}
\description{
This function uses 'ggplot' to plot heatmaps of a network
}
