% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{GeoreferencedPlot.netTools}
\alias{GeoreferencedPlot.netTools}
\title{Plot the given network using its node coordinates}
\usage{
\method{GeoreferencedPlot}{netTools}(
  obj,
  vertex_intensity = "",
  edge_intensity = "",
  xy_axes = TRUE,
  enable_grid = FALSE,
  ...
)
}
\arguments{
\item{obj}{netTools object -> list(graph: igraph, distances_mtx: distances matrix))}
}
\description{
Plot the given network using its node coordinates
}
