% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{GetEventCorrection}
\alias{GetEventCorrection}
\alias{GetEventCorrection.intensitynet}
\title{Gives the event correction value related to the intensitynet object}
\usage{
GetEventCorrection(obj)

\method{GetEventCorrection}{intensitynet}(obj)
}
\arguments{
\item{obj}{intensitynet object}
}
\value{
integer, event correction value
}
\description{
Gives the event correction value related to the intensitynet object
}
\examples{

data("und_intnet_chicago")
GetEventCorrection(und_intnet_chicago)

}
