% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{GetGraph}
\alias{GetGraph}
\alias{GetGraph.intensitynet}
\title{Gives the graph related to the intensitynet object}
\usage{
GetGraph(obj)

\method{GetGraph}{intensitynet}(obj)
}
\arguments{
\item{obj}{intensitynet object}
}
\value{
igraph class object
}
\description{
Returns the 'igraph' class network related to the intensitynet object
}
\examples{

data("und_intnet_chicago")
GetGraph(und_intnet_chicago)

}
