% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{GetGraphType}
\alias{GetGraphType}
\alias{GetGraphType.intensitynet}
\title{Gives the type of graph related to the intensitynet object}
\usage{
GetGraphType(obj)

\method{GetGraphType}{intensitynet}(obj)
}
\arguments{
\item{obj}{intensitynet object}
}
\value{
graph type in characters
}
\description{
Gives the type of graph related to the intensitynet object
}
\examples{

data("und_intnet_chicago")
GetGraphType(und_intnet_chicago)

}
