% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{gplot.intensitynet}
\alias{gplot.intensitynet}
\title{Plot the network and if specified, the moran_i or geary_g heatmap.}
\usage{
\method{gplot}{intensitynet}(obj, intensity = NULL, heatmap = "none", ...)
}
\arguments{
\item{obj}{intensitynet object}

\item{intensity}{vector containing the intensity values that the heatmaps
will use. Default value = NULL}

\item{heatmap}{local 'moran_i' or 'geary_g'}

\item{...}{extra arguments for the class gplot}
}
\description{
Plot the network and if specified, the moran_i or geary_g heatmap.
}
