% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{InitGraph.netTools}
\alias{InitGraph.netTools}
\alias{SetNetCoords.netTools}
\title{Creates an igraph network with the given data}
\usage{
\method{InitGraph}{netTools}(obj)

\method{SetNetCoords}{netTools}(obj)
}
\arguments{
\item{obj}{netTools object -> list(graph: igraph, list(): with the node coordinates 'x' and 'y')}
}
\value{
igraph network

igraph network with the given coordinates as the attributes of the nodes
}
\description{
Creates an igraph network with the given data

Set igraph network node coordinates as its attributes
}
