\name{intensitynet-package}
\alias{intensitynet-package}
\alias{intensitynet}
\docType{package}
\title{
\packageTitle{intensitynet}
}
\description{
\packageDescription{intensitynet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{intensitynet}
\packageIndices{intensitynet}

}
\author{
\packageAuthor{intensitynet}

Maintainer: \packageMaintainer{intensitynet}
}
\references{
Eckardt, M and Mateu, J (2016). Point patterns occurring on complex structures in space and space-time: An alternative network approach.}   
