% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{IsIntensitynet}
\alias{IsIntensitynet}
\alias{summary.intensitynet}
\title{Summary of the intensitynet object}
\usage{
\method{summary}{intensitynet}(obj)

IsIntensitynet(obj)
}
\arguments{
\item{obj}{}
}
\value{
boolean, 'TRUE' if the argument obj is a intensitynet object
}
\description{
Give information about TODO: Finish

Determine if the given object is from the class intensitynet
}
\examples{

data("und_intnet_chicago")
summary(und_intnet_chicago)


data("und_intnet_chicago")
IsIntensitynet(und_intnet_chicago)

}
