% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabDat.R
\name{LabDat}
\alias{LabDat}
\title{Calculation of edge- and nodewise means for planar networks}
\usage{
LabDat(adj, x.coord.node, y.coord.node, x.event, y.event, codats)
}
\arguments{
\item{adj}{A adjacency matrix encoding network structure}

\item{x.coord.node}{X-coordinates of node in network}

\item{y.coord.node}{Y-coordinates of node in network}

\item{x.event}{X-coordinates of event}

\item{y.event}{Y-coordinates of event}

\item{codats}{}
}
\value{
dats vector containing the edgewise means

datmeans vector containing the nodewise means per neighbours
}
\description{
Calculates edgewise and nodewise means of covariates for planar networks
}
\author{
Matthias Eckardt
}

