% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netintensity.R
\name{netintensity}
\alias{netintensity}
\title{Intensity Calculation over networks}
\usage{
netintensity(adj, x.coord.node, y.coord.node, x.event, y.event)
}
\arguments{
\item{adj}{A adjacency matrix encoding network structure}

\item{x.coord.node}{X-coordinates of node in network}

\item{y.coord.node}{Y-coordinates of node in network}

\item{x.event}{X-coordinates of event}

\item{y.event}{Y-coordinates of event}
}
\value{
counts vector containing the mean.intensity per node

edge.counts array containing edge.intensity for adjacent edges

g underlying graph object

deg Vector of degrees per node
}
\description{
Calculates edgewise and mean nodewise intensity fucntion for undirected networks
}
\author{
Matthias Eckardt
}

