% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{NodeGeneralCorrelation.intensitynet}
\alias{NodeGeneralCorrelation.intensitynet}
\title{Gives general node correlation of the network (choosing from: normal correlation, covariance, 
moran-i or geary)}
\usage{
\method{NodeGeneralCorrelation}{intensitynet}(obj, dep_type, lag_max, intensity)
}
\arguments{
\item{obj}{intensitynet object}

\item{dep_type}{the type of dependence statistic to be computed ("correlation", "covariance",
"moran", "geary").}

\item{lag_max}{Maximum geodesic lag at which to compute dependence}

\item{intensity}{vector containing the intensity values that the heatmaps}
}
\value{
A vector containing the dependence statistics (ascending from order 0).
}
\description{
Gives general node correlation of the network (choosing from: normal correlation, covariance, 
moran-i or geary)
}
