% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetDir.R
\name{nodeIntensity.intensitynetDir}
\alias{nodeIntensity.intensitynetDir}
\alias{MeanNodeIntensity.intensitynetDir}
\title{Given a node, calculates its mean intensities regarding in and out edges associated with the node.}
\usage{
\method{MeanNodeIntensity}{intensitynetDir}(obj, node_id)
}
\arguments{
\item{obj}{intensitynetDir object}

\item{node_id}{ID of the node}
}
\value{
mean intensities of the given node for in and out edges
}
\description{
Given a node, calculates its mean intensities regarding in and out edges associated with the node.
}
