% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetMix.R
\name{nodeIntensity.intensitynetMix}
\alias{nodeIntensity.intensitynetMix}
\alias{MeanNodeIntensity.intensitynetMix}
\title{Given a node, calculates its mean intensites depending of the edges associated with the node, those intensities are: 
in, out (for directed edges), undirected and total intensity.}
\usage{
\method{MeanNodeIntensity}{intensitynetMix}(obj, node_id)
}
\arguments{
\item{obj}{intensitynetMix object}

\item{node_id}{ID of the node}
}
\value{
mean intensities of the given node for undirected edges, in and out directed and total intensity.
}
\description{
Given a node, calculates its mean intensites depending of the edges associated with the node, those intensities are: 
in, out (for directed edges), undirected and total intensity.
}
