% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetUnd.R
\name{nodeIntensity.intensitynetUnd}
\alias{nodeIntensity.intensitynetUnd}
\alias{MeanNodeIntensity.intensitynetUnd}
\title{Calculates the mean intensity of the given node (for undirected networks)}
\usage{
\method{MeanNodeIntensity}{intensitynetUnd}(obj, node_id)
}
\arguments{
\item{obj}{intensitynetUnd object}

\item{node_id}{ID of the node}
}
\value{
mean intensity of the given node
}
\description{
Calculates the mean intensity of the given node (intensity of all the edges of the node/number of edges of the node)
}
