% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Nodes of traffic network for Castellon}
\format{An object of class \code{data.frame} with 1611 rows and 2 columns.}
\usage{
data(nodes)
}
\description{
Data provides the coordinates for all nodes for traffic net of Castellon
}
\references{
Eckardt, M and Mateu, J (2016). Point patterns occurring on complex structures in space and space-time: An alternative network approach.
}
\keyword{datasets}

