% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{PathIntensity}
\alias{PathIntensity}
\title{Calculates the intensity of the given path}
\usage{
PathIntensity(obj, path_nodes)
}
\arguments{
\item{obj}{intensitynet object}

\item{path_nodes}{vector containing the node ID's of the path}
}
\value{
intensity of the path
}
\description{
Calculates the intensity of the given path
}
\examples{
\dontrun{
# intnet_chicago -> is an intensitynet object, check example from 'intensitynet' function
short_path <- ShortestPathIntensity(intnet_chicago, node_id1 = 'V1', node_id2 = 'V300')
PathIntensity(intnet_chicago, short_dist$path)
}
}
