% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetDir.R
\name{plot.intensitynetDir}
\alias{plot.intensitynetDir}
\title{Plot intensitynet object}
\usage{
\method{plot}{intensitynetDir}(
  obj,
  vertex_intensity = "none",
  edge_intensity = "none",
  xy_axes = TRUE,
  enable_grid = FALSE,
  ...
)
}
\arguments{
\item{obj}{intensitynet object}
}
\description{
Plot intensitynet object
}
