% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetMix.R
\name{plot.intensitynetMix}
\alias{plot.intensitynetMix}
\title{Plot intensitynet object}
\usage{
\method{plot}{intensitynetMix}(
  obj,
  vertex_intensity = "none",
  edge_intensity = "none",
  xy_axes = TRUE,
  enable_grid = FALSE,
  ...
)
}
\arguments{
\item{obj}{intensitynet object}
}
\description{
Plot intensitynet object
}
