% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{PlotNeighborhood.intensitynet}
\alias{PlotNeighborhood.intensitynet}
\title{Plot the net and the events in the neighborhood area of the given node}
\usage{
\method{PlotNeighborhood}{intensitynet}(obj, node_id, ...)
}
\arguments{
\item{obj}{Intensitynet object}

\item{node_id}{Id of the node which the plot will be focused}

\item{...}{Extra arguments for plotting}
}
\description{
Plot the net and the events in the neighborhood area of the given node
}
\examples{
\dontrun{
# intnet_chicago -> is an intensitynet object, check example from 'intensitynet' function
PlotNeighborhood(intnet_chicago, node_id = 'V300')
}
}
