% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{PlotNeighborhood}
\alias{PlotNeighborhood}
\title{Plot the net and the events in the neighborhood area of the given node}
\usage{
PlotNeighborhood(obj, node_id, ...)
}
\arguments{
\item{obj}{Intensitynet object}

\item{node_id}{Id of the node which the plot will be focused}

\item{...}{Extra arguments for plotting}
}
\value{
No return value, just plots the neighborhood and the events.
}
\description{
Plot the net and the events in the neighborhood area of the given node
}
\examples{

data("und_intnet_chicago")
PlotNeighborhood(und_intnet_chicago, node_id = 'V300')

}
