% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{PointToLine.netTools}
\alias{PointToLine.netTools}
\title{Gives the distance between an event and the line formed by two nodes.}
\usage{
\method{PointToLine}{netTools}(obj)
}
\arguments{
\item{obj}{netTools object -> list(p1:c(coordx, coordy), p2:c(coordx, coordy), e:c(coordx, coordy))}
}
\value{
the distance to the line
}
\description{
Gives the distance between an event and the line (not segment) formed by two nodes.
}
