% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{PointToSegment.netTools}
\alias{PointToSegment.netTools}
\alias{PointToSegment}
\title{Gives the shortest distance between an event and a set of segments.}
\usage{
PointToSegment(obj)
}
\arguments{
\item{obj}{netTools object -> list(p1:matrix(coordx, coordy), p2:matrix(coordx, coordy), e:matrix(coordx, coordy))}
}
\value{
distance vector to each segment
}
\description{
Gives the shortest distance between an event and a set of segments.
}
