% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetMix.R
\name{RelateEventsToNetwork.intensitynetMix}
\alias{RelateEventsToNetwork.intensitynetMix}
\title{Calculates intensity statistics for the given intensitynet object}
\usage{
\method{RelateEventsToNetwork}{intensitynetMix}(obj)
}
\arguments{
\item{obj}{intensitynetMix object}
}
\value{
proper intensitynetMix object with a graph containing the nodewise intensity in the node 
attributes and the edgewise intensities and event covariate proportions as edge attributes.
}
\description{
Calculates edgewise and mean nodewise intensities for Mixed networks and, for each edge, the proportions of
all event covariates.
}
