% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetUnd.R
\name{RelateEventsToNetwork.intensitynetUnd}
\alias{RelateEventsToNetwork.intensitynetUnd}
\title{Calculates intensity statistics for the given intensitynet object}
\usage{
\method{RelateEventsToNetwork}{intensitynetUnd}(obj)
}
\arguments{
\item{obj}{intensitynetUnd object}
}
\value{
proper intensitynetUnd object with a graph containing the nodewise intensity in the node 
attributes and the edgewise intensities and event covariate proportions as edge attributes.
}
\description{
Calculates edgewise and mean nodewise intensities for the given intensitynet object and, for each edge, the proportions of
all event covariates.
}
