% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{RelateEventsToNetwork}
\alias{RelateEventsToNetwork}
\title{Calculates edgewise and mean nodewise intensities for the given intensitynet object and, for each edge, the proportions of
all event covariates.}
\usage{
RelateEventsToNetwork(obj)
}
\arguments{
\item{obj}{intensitynet object}
}
\value{
proper intensitynet object (Undirected, Directed, or Mixed) with a graph containing the nodewise intensity in the node 
attributes and the edgewise intensities and event covariate proportions as edge attributes.
}
\description{
Calculates edgewise and mean nodewise intensities for the given intensitynet object and, for each edge, the proportions of
all event covariates.
}
\examples{

data("und_intnet_chicago")
intnet_chicago <- RelateEventsToNetwork(und_intnet_chicago)

}
