% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{SetNodeIntensity.netTools}
\alias{SetNodeIntensity.netTools}
\title{Sets the given intensities as a node attribute to the given igraph network}
\usage{
\method{SetNodeIntensity}{netTools}(obj)
}
\arguments{
\item{obj}{netTools object -> list(graph: igraph, node_id: node id, intensity: node intensity)}
}
\value{
igraph network with the given intensities as attributes of the nodes
}
\description{
Sets the given intensities as a node attribute to the given igraph network
}
