% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{ShortestDistance.netTools}
\alias{ShortestDistance.netTools}
\title{Calculates the shortest distance path between two nodes}
\usage{
\method{ShortestDistance}{netTools}(obj)
}
\arguments{
\item{obj}{netTools object -> list(graph: igraph, node_id1: node id, node_id2: node id, distances_mtx: distances matrix))}
}
\value{
distance of the path and the nodes of the path
}
\description{
Calculates the shortest distance path between two nodes
}
