% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{ShortestNodeDistance.intensitynet}
\alias{ShortestNodeDistance.intensitynet}
\title{Calculates the shortest distance path between two nodes (based on the minimum amount of edges).
The function also returns the total weight of the path, if the weight is not available, returns
the number of edges.}
\usage{
\method{ShortestNodeDistance}{intensitynet}(obj, node_id1, node_id2)
}
\arguments{
\item{obj}{intensitynet object}

\item{node_id1}{id of the starting node}

\item{node_id2}{id of the end node}
}
\value{
distance of the path and the nodes of the path
}
\description{
Calculates the shortest distance path between two nodes (based on the minimum amount of edges).
The function also returns the total weight of the path, if the weight is not available, returns
the number of edges.
}
