% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{summary}
\alias{summary}
\alias{summary.intensitynet}
\title{Summary of the intensitynet object}
\usage{
\method{summary}{intensitynet}(object, ...)
}
\arguments{
\item{object}{Intensitynet object}

\item{...}{Extra parameters for the summary function}
}
\value{
list containing the displayed information
}
\description{
Give information about the intensytinet object specific class (intensitynetUnd, intensitynetDir,
 or intensitynetMix), the network number of nodes, edges and events, the event correction value and,
 if the events had been related to the intensitynet object network.
}
\examples{

data("und_intnet_chicago")
summary(und_intnet_chicago)

}
