% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{Undirected2RandomDirectedAdjMtx.netTools}
\alias{Undirected2RandomDirectedAdjMtx.netTools}
\title{Creates a directed adjacency matrix from an Undirected one with random directions (in-out edges) 
but with the same connections between nodes.}
\usage{
\method{Undirected2RandomDirectedAdjMtx}{netTools}(obj)
}
\arguments{
\item{obj}{netTools object -> list(mtx: matrix)}
}
\value{
directed adjacency matrix with random directions
}
\description{
Creates a directed adjacency matrix from an Undirected one with random directions (in-out edges) 
but with the same connections between nodes.
}
