% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{ApplyWindow.intensitynet}
\alias{ApplyWindow.intensitynet}
\title{Get the intensitynet object delimited by the given window}
\usage{
\method{ApplyWindow}{intensitynet}(obj, x_coords, y_coords)
}
\arguments{
\item{obj}{intensitynet object}

\item{x_coords}{vector containing the x coordinate limits of the window}

\item{y_coords}{vector containing the y coordinate limits of the window}
}
\value{
intensitynet object delimited by the window (sub-part of the original)
}
\description{
Get the intensitynet object delimited by the given window
}
\examples{

data("und_intnet_chicago")
sub_intnet_chicago <- ApplyWindow(und_intnet_chicago, 
                                  x_coords = c(300, 900), 
                                  y_coords = c(500, 1000))

}
