% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{CalculateEventIntensities}
\alias{CalculateEventIntensities}
\title{Calculates edgewise and mean nodewise intensities for the given intensitynet object}
\usage{
CalculateEventIntensities(obj)
}
\arguments{
\item{obj}{intensitynetDir object}
}
\value{
proper intensitynet object (Undirected, Directed or Mixed) with a graph containing all the intensities as 
attributes of its nodes and edges
}
\description{
Calculates edgewise and mean nodewise intensities for the given intensitynet object
}
\examples{

data("und_intnet_chicago")
intnet_chicago <- CalculateEventIntensities(und_intnet_chicago)

}
