% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{EdgeIntensity.intensitynet}
\alias{EdgeIntensity.intensitynet}
\title{If not calculated, calculates the intensity of the edge with nodes; node_id1, node_id2. 
If the edge already contains an intensity, give it directly.}
\usage{
\method{EdgeIntensity}{intensitynet}(obj, node_id1, node_id2, z = 5)
}
\arguments{
\item{obj}{intensitynet object}

\item{node_id1}{First node ID of the edge}

\item{node_id2}{Second node ID of the edge}

\item{z}{Maximum distance between the event and the edge to consider the event part of the edge.}
}
\value{
Intensity of the edge
}
\description{
If not calculated, calculates the intensity of the edge with nodes; node_id1, node_id2. 
If the edge already contains an intensity, give it directly.
}
