% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{GeoreferencedGgplot2.netTools}
\alias{GeoreferencedGgplot2.netTools}
\title{This function uses 'ggplot' to plot heatmaps of a network}
\usage{
\method{GeoreferencedGgplot2}{netTools}(obj, ...)
}
\arguments{
\item{obj}{netTools object -> list(graph: igraph, data_df: dataframe(intensity: intensity of the nodes, 
xcoord: x coordinates of the nodes, ycoord: y coordinates of the nodes, heattype: data which the heatmap will refer), 
mode: ('moran', 'getis' or 'intensity'))}

\item{...}{extra arguments for the ggplot}
}
\description{
This function uses 'ggplot' to plot heatmaps of a network
}
