% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{GeoreferencedPlot.netTools}
\alias{GeoreferencedPlot.netTools}
\title{Plot the given network using its node coordinates}
\usage{
\method{GeoreferencedPlot}{netTools}(obj, ...)
}
\arguments{
\item{obj}{netTools object -> list(
intnet: intensitynet object, 
vertex_labels: list of labels for the vertices,
edge_labels: list of labels for the edges, 
xy_axes: boolean to show or not the x and y axes, 
enable_grid: boolean to draw or not a background grid, show_events: boolean to show or not the events as orange squares,
show_events option to show the events as orange squares, FALSE by default,
alpha optional argument to set the transparency of the events (show_events = TRUE). The range is from 0.1 (transparent) to 1 (opaque). Default: alpha = 1,
path: vector with the nodes of the path to be highlighted. Default NULL)}

\item{...}{extra arguments for the plot}
}
\description{
Plot the given network using its node coordinates
}
