% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{intensitynet}
\alias{intensitynet}
\title{Constructor of the class intensitynet}
\usage{
intensitynet(adjacency_mtx, node_coords, events_mtx, graph_type = "undirected")
}
\arguments{
\item{adjacency_mtx}{Network adjacency matrix}

\item{node_coords}{Nodes latitude and longitude matrix}

\item{events_mtx}{Events latitude and longitude matrix}

\item{graph_type}{Network type: 'undirected' (default), 'directed' or 'mixed'}
}
\value{
intensitynet object containing: graph=<igraph>, events = <matrix>, graph_type = c('directed', 'undirected', 'mixed'), 
distances = <matrix>
}
\description{
Constructor of the class intensitynet
}
