% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{IsIntensitynet}
\alias{IsIntensitynet}
\title{Is this class object intensitynet?}
\usage{
IsIntensitynet(obj)
}
\arguments{
\item{obj}{The object which will be checked if it belongs to the intensitynet class}
}
\value{
boolean, 'TRUE' if the argument obj is a intensitynet object
}
\description{
Determine if the given object is from the class intensitynet
}
\examples{

data("und_intnet_chicago")
IsIntensitynet(und_intnet_chicago)

}
