% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{NodeLocalCorrelation.intensitynet}
\alias{NodeLocalCorrelation.intensitynet}
\title{Gives node local moran-i or geary-c correlations}
\source{
*Luc Anselin. A Local Indicator of Multivariate SpatialAssociation: Extending Geary's c, Geographical Analysis 2018; doi: https://doi.org/10.1111/gean.12164
}
\usage{
\method{NodeLocalCorrelation}{intensitynet}(obj, dep_type = "moran", intensity)
}
\arguments{
\item{obj}{intensitynet object}

\item{dep_type}{'moran', 'getis' or 'geary'. Type of local correlation to be computed (Moran-i, Getis-Gstar, Geary-c*),
default = 'moran'.}

\item{intensity}{vector containing the intensity values that which are used to calculate the correlation.}
}
\value{
An intnet object which contains an igraph network with the selected correlation 
added into the vertices attributes
}
\description{
Gives node local moran-i or geary-c correlations
}
\examples{

# intnet_chicago -> is an intensitynet object, check example from 'intensitynet' function
g <- intnet_chicago$graph
data_moran <- NodeLocalCorrelation(intnet_chicago, 
                                   dep_type = 'moran', 
                                   intensity = igraph::vertex_attr(g)$intensity)
moran_i <- data_moran$correlation
intnet <- data_moran$intnet

}
