% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{NodeLocalCorrelation.intensitynet}
\alias{NodeLocalCorrelation.intensitynet}
\title{Gives node local moran-i or geary-g correlations}
\usage{
\method{NodeLocalCorrelation}{intensitynet}(obj, dep_type = "moran_i", intensity)
}
\arguments{
\item{obj}{intensitynet object}

\item{dep_type}{the type of dependence statistic to be computed ('moran_i' or 'geary_g'),
default = 'moran_i.}

\item{intensity}{vector containing the intensity values that the heatmaps}
}
\value{
An intnet object wich contains a igraph network with the selected correlation added
}
\description{
Gives node local moran-i or geary-g correlations
}
