% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{NodeLocalCorrelation}
\alias{NodeLocalCorrelation}
\title{Gives node local moran-i or geary-c correlations}
\source{
*"A Local Indicator of Multivariate SpatialAssociation: Extending Geary's c, Geographical Analysis" Luc Anselin (2018) <doi:10.1111/gean.12164>
}
\usage{
NodeLocalCorrelation(obj, dep_type = "moran", intensity)
}
\arguments{
\item{obj}{intensitynet object}

\item{dep_type}{'moran', 'getis' or 'geary'. Type of local correlation to be computed (Moran-i, Getis-Gstar, Geary-c*),
default = 'moran'.}

\item{intensity}{vector containing the intensity values that which are used to calculate the correlation.}
}
\value{
intensitynet class object which contains an igraph network with the selected correlation 
added into the vertices attributes
}
\description{
Gives node local moran-i or geary-c correlations
}
\examples{
\dontrun{
data("und_intnet_chicago")
g <- und_intnet_chicago$graph
data_moran <- NodeLocalCorrelation(und_intnet_chicago, 
                                   dep_type = 'moran', 
                                   intensity = igraph::vertex_attr(g)$intensity)
moran_i <- data_moran$correlation
intnet <- data_moran$intnet
}
}
