% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{PathIntensity.intensitynet}
\alias{PathIntensity.intensitynet}
\title{Calculates the intensity of the given path}
\usage{
\method{PathIntensity}{intensitynet}(obj, path_nodes)
}
\arguments{
\item{obj}{intensitynet object}

\item{path_nodes}{vector containing the node ID's of the path}
}
\value{
intensity of the path
}
\description{
Calculates the intensity of the given path
}
