% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{PathIntensity}
\alias{PathIntensity}
\title{Calculates the intensity of the given path}
\usage{
PathIntensity(obj, path_nodes)
}
\arguments{
\item{obj}{intensitynet object}

\item{path_nodes}{vector containing the node ID's of the path}
}
\value{
intensity of the path
}
\description{
Calculates the intensity of the given path
}
\examples{

data("und_intnet_chicago")
PathIntensity(und_intnet_chicago, c('V115', 'V123', 'V125', 'V134'))

}
