% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetDir.R
\name{plot.intensitynetDir}
\alias{plot.intensitynetDir}
\title{Plot intensitynet object}
\usage{
\method{plot}{intensitynetDir}(
  x,
  vertex_labels = "none",
  edge_labels = "none",
  xy_axes = TRUE,
  enable_grid = FALSE,
  ...
)
}
\arguments{
\item{x}{intensitynet object}

\item{vertex_labels}{list -> labels for the vertices}

\item{edge_labels}{list -> labels for the edges}

\item{xy_axes}{show the x and y axes}

\item{enable_grid}{draw a background grid}

\item{...}{extra arguments for the plot}
}
\description{
Plot intensitynet object
}
