% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynetMix.R
\name{plot.intensitynetMix}
\alias{plot.intensitynetMix}
\title{Plot intensitynet object}
\usage{
\method{plot}{intensitynetMix}(
  x,
  vertex_labels = "none",
  edge_labels = "none",
  xy_axes = TRUE,
  enable_grid = FALSE,
  show_events = FALSE,
  ...
)
}
\arguments{
\item{x}{intensitynet object}

\item{vertex_labels}{list -> labels for the vertices}

\item{edge_labels}{list -> labels for the edges}

\item{xy_axes}{show the x and y axes}

\item{enable_grid}{draw a background grid}

\item{show_events}{option to show the events as orange squares, FALSE by default}

\item{...}{extra arguments for the plot}
}
\value{
No return value, same as graphics::plot.
}
\description{
Plot intensitynet object
}
\examples{

data("mix_intnet_chicago")
plot(mix_intnet_chicago) # basic plot
plot(mix_intnet_chicago, enable_grid = TRUE) # with grid
plot(mix_intnet_chicago, xy_axes = FALSE) # without axes

}
