% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{PointToSegment.netTools}
\alias{PointToSegment.netTools}
\alias{PointToSegment}
\title{Return the shortest distance between an event and the segment formed by two nodes.}
\usage{
PointToSegment(obj)
}
\arguments{
\item{obj}{netTools object -> list(p1:c(coordx, coordy), p2:c(coordx, coordy), e:c(coordx, coordy))}
}
\value{
distance to the segment
}
\description{
Return the shortest distance between an event and the segment formed by two nodes.
}
