% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netTools.R
\name{PointToSegment_deprecated.netTools}
\alias{PointToSegment_deprecated.netTools}
\alias{PointToSegment_deprecated}
\title{Gives the shortest distance between an event and the segment formed by two nodes.}
\usage{
PointToSegment_deprecated(obj)
}
\arguments{
\item{obj}{netTools object -> list(p1:c(coordx, coordy), p2:c(coordx, coordy), e:c(coordx, coordy))}
}
\value{
distance to the segment
}
\description{
Gives the shortest distance between an event and the segment formed by two nodes.
}
