% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensitynet.R
\name{ShortestNodeDistance.intensitynet}
\alias{ShortestNodeDistance.intensitynet}
\title{Calculates the shortest distance path between two nodes}
\usage{
\method{ShortestNodeDistance}{intensitynet}(obj, node_id1, node_id2)
}
\arguments{
\item{obj}{intensitynet object}

\item{node_id1}{id of the starting node}

\item{node_id2}{id of the end node}
}
\value{
distance of the path and the nodes of the path
}
\description{
Calculates the shortest distance path between two nodes
}
